Added FSM, Sprites, and Editor
The new system adds a Finite State Machine to handle different game
states, in order to effectively switch between different modes. The new
Sprite system adds an object oriented method of creating sprites, and
will likely be a base class for future animated and text based sprites/
The world editor is likely to undergo serious revisions, but currently
has a base functionality of placing and removing two types of tiles.